[#1554] Rename lib/onetime/classmethods.rb
