+++ lib/onetime/class_methods.rb
